refactor(components): transformed const to function
