+++ src/components/SequencesCollapse/SequencesCollapse.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
+import './SequencesCollapse.css';
+
 import { Collapse } from 'antd';
 import { CaretRightOutlined } from '@ant-design/icons';
-import SequenceCollapse from '../SequenceCollapse/SequenceCollapse';
 
-const SequencesCollapse: React.FC = () => {
+import { SequenceCollapse } from '../index';
+
+export function SequencesCollapse() {
   const { Panel } = Collapse;
 
   const sequencesName = [
@@ -38,11 +41,8 @@ const SequencesCollapse: React.FC = () => {
       bordered={false}
       defaultActiveKey={['1']}
       expandIcon={({ isActive }) => <CaretRightOutlined rotate={isActive ? 90 : 0} />}
-      className="site-collapse-custom-collapse"
-    >
+      className="site-collapse-custom-collapse">
       {sequences}
     </Collapse>
   );
-};
-
-export default SequencesCollapse;
+}
