+++ src/components/index.ts
@@ -3,3 +3,5 @@ export { NavBar } from './NavBar/NavBar';
 export { NavBarApp } from './NavBarApp/NavBarApp';
 export { ProjectCard } from './ProjectCard/ProjectCard';
 export { RectangularBtn } from './Rectangular-Btn/Rectangular-Btn';
+export { SequenceCollapse } from './SequenceCollapse/SequenceCollapse';
+export { SequencesCollapse } from './SequencesCollapse/SequencesCollapse';
